requirements.txt
Flask==1.1.2
numpy==1.18.2
pandas==1.0.3
opencv-python==4.2.0.32
matplotlib==3.2.1
Pillow==7.1.2
scipy==1.4.1
tabulate==0.8.7
scikit-image==0.17.2
scikit-learn==0.22.2.post1
geojson==2.5.0
Shapely==1.7.0
rdp==0.8
seaborn==0.10.0
tensorflow==1.15.0
h5py==2.10.0
tqdm==4.48.2